requirements.txt
Flask==2.3.2
redis==5.0.0
mysql-connector-python==8.0.33
